Move Ticker class to utils module
