+++ mapek_framework/src/mapek_framework/__init__.py
@@ -3,4 +3,3 @@ from .group import Group
 from .interaction import Interaction
 from .managed_system import ManagedSystem
 from .mape_element import MonitorElement, AnalyzeElement, PlanElement, ExecuteElement
-from .ticker import Ticker
