+++ mapek_framework/src/test_mapek_framework/run_tests.py
@@ -5,7 +5,7 @@ from .test_group import GroupTestCase
 from .test_interaction import InteractionTestCase
 from .test_managed_system import ManagedSystemTestCase
 from .test_mape_element import MapeElementTestCase
-from .test_ticker import TickerTestCase
+from .test_utils import UtilsTestCase
 
 
 if __name__ == '__main__':
@@ -18,4 +18,4 @@ if __name__ == '__main__':
     rosunit.unitrun('test_mapek_framework',
                     'test_mape_element_mapek_framework', MapeElementTestCase)
     rosunit.unitrun('test_mapek_framework',
-                    'test_ticker_mapek_framework', TickerTestCase)
+                    'test_utils_mapek_framework', UtilsTestCase)
